updated generation

+++ src/main/java/by/vstu/electronicjournal/controller/JournalSiteController.java
@@ -32,7 +32,7 @@ public class JournalSiteController {
     @GetMapping("filter")
     public JournalSiteDTO getFilteredByTeacherAndGroupAndDisciplineAndTypeClassAndSubGroup(@RequestParam("teacher_id") Long teacherId,
                @RequestParam("group_name") String groupName, @RequestParam("discipline_id") Long disciplineId, @RequestParam("type_class_id") Long typeClassId,
-               @RequestParam("sub_group_number") Byte subGroupNumber) {
+               @RequestParam("sub_group_number") Integer subGroupNumber) {
         return journalSiteService.getFilteredByTeacherAndGroupAndDisciplineTypeClassAndSubGroup(teacherId, groupName, disciplineId, typeClassId, subGroupNumber);
     }
